requirements.txt
pandas==1.1.3
SQLAlchemy==1.3.20
PyMySQL==1.0.2
Flask==1.1.2
numpy==1.19.2
mysql_connector_repackaged==0.3.1
python_bcrypt==0.3.2
